Small optimizations
that also means, that a new cycle to a more clear now direction can
have its starting point, tagged as 09.009.

Now!

This (to|in (the borders of ((futileness (if such a word exists)) && pain in vain)))

so, this coding project (which it's circle has been logged perfectly in|by the GitHub),
which it's practical means (is described above),
and because is close to close,
is dedicated to a little friend
of mine, which I
		(this humble arrogant bastard human being (which can be a real (b|p) && pain in the a))
I, this human being, had the privilige to live with him,
and I had to kill him, in an out of sudden accident (hitted by a car (who had no responsibility))
(and because (I didn't had the luxury to have a gun),
with a ...,
the god blessed me to diy in|by my hands,
at the same same time when an owl was crawling from the forest

this being had a name, was black back jack, but sounded as|with back,

To the Black Back Jack, : I see you many times during the day, I miss you.